# service/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/user/createAndUpdateRole":{"post":{"operationId":"UserController_createAndUpdateRole","summary":"分配用户角色","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAndUpdateRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/UserEntity"}}}]}}}}},"tags":["用户模块"]}},"/api/user/update":{"post":{"operationId":"UserController_update","summary":"修改用户信息","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/UserEntity"}}}]}}}}},"tags":["用户模块"]}},"/api/user/updateStatus":{"post":{"operationId":"UserController_updateStatus","summary":"修改用户状态","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStatusDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/UserEntity"}}}]}}}}},"tags":["用户模块"]}},"/api/user/updatePassword":{"post":{"operationId":"UserController_updatePassword","summary":"修改用户密码","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/UserEntity"}}}]}}}}},"tags":["用户模块"]}},"/api/user/getUserInfo":{"get":{"operationId":"UserController_getUserInfo","summary":"根据用户ID查询信息","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/UserEntity"}}}]}}}}},"tags":["用户模块"]}},"/api/user/findUserList":{"post":{"operationId":"UserController_findUser","summary":"查询用户列表","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindUserListDto"}}}},"responses":{"201":{"description":""}},"tags":["用户模块"]}},"/api/register":{"post":{"operationId":"BaseController_register","summary":"用户注册","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/UserEntity"}}}]}}}}},"tags":["登录与注册"]}},"/api/login":{"post":{"operationId":"BaseController_login","summary":"用户登录","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/CreateTokenDto"}}}]}}}}},"tags":["登录与注册"]}},"/api/role/createRole":{"post":{"operationId":"RoleController_create","summary":"创建角色","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/RoleEntity"}}}]}}}}},"tags":["角色模块"],"security":[{"bearer":[]}]}},"/api/role/assignPermissions":{"post":{"operationId":"RoleController_assignPermissions","summary":"分配角色权限","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignPermissionsDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/RoleEntity"}}}]}}}}},"tags":["角色模块"],"security":[{"bearer":[]}]}},"/api/role/findRoleMenu":{"post":{"operationId":"RoleController_findRoleMenu","summary":"查询角色菜单权限","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindRoleMenuDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/RoleEntity"}}}]}}}}},"tags":["角色模块"],"security":[{"bearer":[]}]}},"/api/role/deleteRole":{"post":{"operationId":"RoleController_delete","summary":"删除角色","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"type":"null","default":null}}}]}}}}},"tags":["角色模块"],"security":[{"bearer":[]}]}},"/api/role/updateRole":{"post":{"operationId":"RoleController_update","summary":"更改角色信息","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/RoleEntity"}}}]}}}}},"tags":["角色模块"],"security":[{"bearer":[]}]}},"/api/role/findRoleList":{"post":{"operationId":"RoleController_findList","summary":"查询角色列表","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FindRoleListDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/RoleEntity"}}}}]}}}}},"tags":["角色模块"],"security":[{"bearer":[]}]}},"/api/common/getRoleEnum":{"get":{"operationId":"CommonController_getRoleEnum","summary":"获取角色枚举","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"type":"null","default":null}}}]}}}}},"tags":["公告模块"],"security":[{"bearer":[]}]}},"/api/menu/createMenu":{"post":{"operationId":"MenuController_create","summary":"创建菜单","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMenuDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/MenuEntity"}}}]}}}}},"tags":["菜单模块"]}},"/api/menu/updateMenu":{"post":{"operationId":"MenuController_update","summary":"创建菜单","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMenuDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/MenuEntity"}}}]}}}}},"tags":["菜单模块"]}},"/api/menu/findMenuList":{"post":{"operationId":"MenuController_findMenuList","summary":"查询菜单列表","parameters":[{"name":"Authorization","in":"header","description":"本次请求请带上token","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/ResultData"},{"properties":{"data":{"$ref":"#/components/schemas/MenuEntity"}}}]}}}}},"tags":["菜单模块"]}}},"info":{"title":"博客管理系统API文档","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateAndUpdateRoleDto":{"type":"object","properties":{"userId":{"type":"string","description":"用户ID"},"roleId":{"type":"string","description":"角色ID"}},"required":["userId","roleId"]},"UpdateUserDto":{"type":"object","properties":{"id":{"type":"string","description":"用户编码"},"status":{"type":"number","description":"所属状态: 1-有效，0-禁用","enum":[0,1]},"phone":{"type":"string","description":"手机号"},"email":{"type":"string","description":"邮箱"},"describe":{"type":"string","description":"描述"},"avatar":{"type":"string","description":"头像"}},"required":["id"]},"UpdateStatusDto":{"type":"object","properties":{"id":{"type":"string","description":"账号ID"},"status":{"type":"number","description":"所属状态: 1-有效，0-禁用","enum":[0,1]}},"required":["id","status"]},"UpdatePasswordDto":{"type":"object","properties":{"id":{"type":"string","description":"账号ID"},"password":{"type":"string","description":"密码"}},"required":["id","password"]},"FindUserListDto":{"type":"object","properties":{"page":{"type":"number","description":"显示页数"},"size":{"type":"number","description":"每页显示条数"},"status":{"type":"number","description":"所属状态: 1-有效，0-禁用","enum":[0,1]},"roleId":{"type":"string","description":"拥有角色id"},"account":{"type":"string","description":"账号模糊搜索"}},"required":["page","size","status"]},"ResultData":{"type":"object","properties":{"code":{"type":"number","default":200},"msg":{"type":"string","default":"请求成功"}},"required":["code","msg"]},"UserEntity":{"type":"object","properties":{"id":{"type":"string","description":"id"},"account":{"type":"string","description":"用户登录账号"},"phone":{"type":"string","description":"手机号"},"email":{"type":"string","description":"邮箱"},"status":{"type":"number","description":"所属状态: 1-有效，0-禁用","enum":[0,1]},"avatar":{"type":"string","description":"头像url"},"describe":{"type":"string","description":"描述"},"createDate":{"format":"date-time","type":"string","description":"创建时间"},"updateDate":{"format":"date-time","type":"string","description":"更新时间"}},"required":["id","account","phone","email","status","avatar","describe","createDate","updateDate"]},"CreateTokenDto":{"type":"object","properties":{"accessToken":{"type":"string","description":"accessToken"},"refreshToken":{"type":"string","description":"refreshToken"}},"required":["accessToken","refreshToken"]},"CreateUserDto":{"type":"object","properties":{"account":{"type":"string","description":"用户账号"},"password":{"type":"string","description":"密码"},"phone":{"type":"string","description":"手机号"},"email":{"type":"string","description":"邮箱"},"confirmPassword":{"type":"string","description":"确认密码"},"describe":{"type":"string","description":"描述"},"avatar":{"type":"string","description":"头像"}},"required":["account","password","confirmPassword"]},"LoginUserDto":{"type":"object","properties":{"account":{"type":"string","description":"账号"},"password":{"type":"string","description":"密码"}},"required":["account","password"]},"RoleEntity":{"type":"object","properties":{"id":{"type":"string","description":"id"},"name":{"type":"string","description":"角色名称"},"remark":{"type":"string","description":"角色备注"},"availableRange":{"type":"string","description":"可用范围"},"createDate":{"format":"date-time","type":"string","description":"创建时间"},"updateDate":{"format":"date-time","type":"string","description":"更新时间"}},"required":["id","name","remark","availableRange","createDate","updateDate"]},"CreateRoleDto":{"type":"object","properties":{"name":{"type":"string","description":"角色名称"},"availableRange":{"type":"number","description":"可用范围: 1-后台管理系统，0-博客页面，2全部","enum":[0,1,2]},"remark":{"type":"string","description":"角色备注"},"menuIds":{"description":"当前角色所拥有的菜单组","type":"array","items":{"type":"string"}}},"required":["name","availableRange","menuIds"]},"AssignPermissionsDto":{"type":"object","properties":{"roleId":{"type":"string","description":"roleId"},"menuIds":{"description":"menuIds","type":"array","items":{"type":"string"}}},"required":["roleId","menuIds"]},"FindRoleMenuDto":{"type":"object","properties":{"roleId":{"type":"string","description":"roleId"}},"required":["roleId"]},"UpdateRoleDto":{"type":"object","properties":{"id":{"type":"string","description":"id"},"name":{"type":"string","description":"角色名称"},"availableRange":{"type":"number","description":"可用范围: 1-后台管理系统，0-博客页面，2全部","enum":[0,1,2]},"remark":{"type":"string","description":"角色备注"},"menuIds":{"description":"当前角色所拥有的菜单组","type":"array","items":{"type":"string"}}},"required":["id","name","availableRange","menuIds"]},"FindRoleListDto":{"type":"object","properties":{"page":{"type":"number","description":"显示页数"},"size":{"type":"number","description":"每页显示条数"},"name":{"type":"string","description":"角色名称模糊搜索"}},"required":["page","size"]},"CreateMenuDto":{"type":"object","properties":{"menuName":{"type":"string","description":"菜单名称"},"remark":{"type":"string","description":"菜单备注"},"menuType":{"type":"string","description":"菜单类型"},"parentId":{"type":"string","description":"上级菜单"},"menuPath":{"type":"string","description":"跳转路径"},"menuStatus":{"type":"string","description":"状态"},"filePath":{"type":"string","description":"路由文件路径"},"sort":{"type":"string","description":"排序"},"key":{"type":"string","description":"键值"},"otherProperties":{"type":"object","description":"其他属性"}}},"UpdateMenuDto":{"type":"object","properties":{"id":{"type":"string","description":"id"},"menuName":{"type":"string","description":"菜单名称"},"remark":{"type":"string","description":"菜单备注"},"menuType":{"type":"string","description":"菜单类型"},"parentId":{"type":"string","description":"上级菜单"},"menuPath":{"type":"string","description":"跳转路径"},"menuStatus":{"type":"string","description":"状态"},"filePath":{"type":"string","description":"路由文件路径"},"sort":{"type":"string","description":"排序"},"key":{"type":"string","description":"键值"},"otherProperties":{"type":"object","description":"其他属性"}},"required":["id"]}}}}